// src/types/v1021.ts
export interface Offer {
    account: AccountId32
    total: bigint
    rate: number
    minAverageRewardRate: bigint
    deposit: bigint
}
// ...
export type AccountId32 = Bytes
// ...
export const Offer: sts.Type<Offer> = sts.struct(() => {
    return {
        account: AccountId32,
        total: sts.bigint(),
        rate: sts.number(),
        minAverageRewardRate: sts.bigint(),
        deposit: sts.bigint(),
    }
})
// ...
export const AccountId32 = sts.bytes()