// src/types/enjinV1023.ts
export interface Offer {
    account: AccountId32
    total: bigint
    rate: number
    minAverageRewardRate: bigint
    deposit: bigint
    tokenFilter: TokenFilter
}

export type TokenFilter = TokenFilter_All | TokenFilter_BlockList | TokenFilter_Whitelist

export interface TokenFilter_All {
    __kind: 'All'
}

export interface TokenFilter_BlockList {
    __kind: 'BlockList'
    value: bigint[]
}

export interface TokenFilter_Whitelist {
    __kind: 'Whitelist'
    value: bigint[]
}

export type AccountId32 = Bytes
// ...
export const TokenFilter: sts.Type<TokenFilter> = sts.closedEnum(() => {
    return {
        All: sts.unit(),
        BlockList: sts.array(() => sts.bigint()),
        Whitelist: sts.array(() => sts.bigint()),
    }
})
// ...
export const AccountId32 = sts.bytes()
// ...
export const Offer: sts.Type<Offer> = sts.struct(() => {
    return {
        account: AccountId32,
        total: sts.bigint(),
        rate: sts.number(),
        minAverageRewardRate: sts.bigint(),
        deposit: sts.bigint(),
        tokenFilter: TokenFilter,
    }
})